Add files via upload
The program is pretty basic nothing fancy, in the future probably I'll come back again to see the messy code I've written here, but it's a normal checkpoint in anybody's learning process.
Based on a list of students the user submits containing various informations about them, the program executes multiple basic functions depending on the user's choice. When the user quits the program properly, the informations get saved in a file.